[feature] add possibility to decide if existing json should be appended or overwritten

+++ imdb-scrapper.py
@@ -95,6 +95,7 @@ if __name__ == '__main__':
     parser.add_argument('storing', choices=['save', 'unsave'],
                         help='[save] store movies data after each request,[unsave] store movies data after all requests were executed')
     parser.add_argument('--start', type=int, help='the movie id to start with')
+    parser.add_argument('--overwrite', default='yes', choices=['yes', 'no'], help='[yes] overwrite json file, [no] append json file')
     args = parser.parse_args()
 
     argcomplete.autocomplete(parser)
@@ -117,13 +118,25 @@ if __name__ == '__main__':
     # Proxy the requests
     imdb = Imdb(anonymize=True)
 
-    # Create a clean json file
-    logger.info('JSON file "movies.json" created.')
-    with open('movies.json', mode='w') as moviesjson:
-        json.dump({'movies': []}, moviesjson)
+    if args.overwrite == 'yes':
+        # Create a clean json file
+        logger.info('JSON file "movies.json" created.')
+        with open('movies.json', mode='w') as moviesjson:
+            json.dump({'movies': []}, moviesjson)
 
-    # Create a dictionary for the movie
-    movies = {'movies': []}
+        # Create a dictionary for the movies
+        movies = {'movies': []}
+    else:
+        # Load the dictionary from json file
+        with open('movies.json') as f:
+            movies_temp = json.load(f)
+
+        # Create a dictionary for the movies
+        movies = {'movies': []}
+
+        # Load data from file into created movies dictionary
+        for i in range(0, len(movies_temp)):
+            movies['movies'].append(movies_temp[i])
 
     # Process N films of IMDb
     logger.info('Movie retrieval started.')
@@ -142,7 +155,7 @@ if __name__ == '__main__':
             gen_json(movies)
             logger.info('Movie with id ' + movie['id'] + ' stored in "movies.json".')
 
-    # Store the update movies dictionary in the json file (after all movies were retrieved)
+    # Store the updated movies dictionary in the json file (after all movies were retrieved)
     if args.storing == 'unsave':
         gen_json(movies)
         logger.info('All retrieved movies were stored in "movies.json".')
